servers/workspace-doc-generator/pyproject.toml
[project]
name = "mcp-workspace-doc-generator"
version = "0.2.0"
description = "MCP server that builds a workspace document from artifacts and returns a cam.asset.file_detail artifact."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Astra Team" }]

dependencies = [
  "mcp[server]>=1.3.0,<2",         # Official MCP SDK with streamable-http + stdio support
  "pydantic>=2.7",
  "pydantic-core>=2.18",
  "orjson>=3.10",
  "PyYAML>=6.0",
  "httpx>=0.27",
  "openai>=1.51.0",
  "uvicorn>=0.30",
  "uvloop>=0.19; platform_system != 'Windows'",
  "click>=8.1",
]

[project.scripts]
mcp-workspace-doc-generator = "mcp_workspace_doc_generator.__main__:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_workspace_doc_generator"]

[tool.uv]
dev-dependencies = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]

[tool.ruff]
line-length = 100